+++ locales/en/plugin__kubevirt-plugin.json
@@ -17,10 +17,9 @@
   "{{availableText}} available": "{{availableText}} available",
   "{{badList}} is not a valid {{searchKey}} value.": "{{badList}} is not a valid {{searchKey}} value.",
   "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.": "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.",
+  "{{clusterCount}} clusters excluded": "{{clusterCount}} clusters excluded",
   "{{clustersCount}} clusters found": "{{clustersCount}} clusters found",
   "{{clustersCount}} clusters, {{projectsCount}} projects found": "{{clustersCount}} clusters, {{projectsCount}} projects found",
-  "{{count}} clusters excluded_one": "{{count}} clusters excluded",
-  "{{count}} clusters excluded_other": "{{count}} clusters excludeds",
   "{{count}} Expression_one": "{{count}} Expression",
   "{{count}} Expression_other": "{{count}} Expressions",
   "{{count}} failed checks_one": "{{count}} failed check",
@@ -398,9 +397,11 @@
   "Clear search": "Clear search",
   "Clear selection": "Clear selection",
   "Click \"Configure features\" to install this feature": "Click \"Configure features\" to install this feature",
+  "Click a cluster to select it as the migration target.": "Click a cluster to select it as the migration target.",
   "Click any example below to try it in the search bar": "Click any example below to try it in the search bar",
   "Click Connect to open serial console.": "Click Connect to open serial console.",
   "Click Connect to open the VNC console.": "Click Connect to open the VNC console.",
+  "Click the cluster to select it as the migration target.": "Click the cluster to select it as the migration target.",
   "Clone": "Clone",
   "Clone {{kind}}": "Clone {{kind}}",
   "Clone {{sourceKind}}": "Clone {{sourceKind}}",
@@ -437,7 +438,7 @@
   "Cluster preferences": "Cluster preferences",
   "Cluster Preferences": "Cluster Preferences",
   "Cluster preferences table": "Cluster preferences table",
-  "Cluster recommendations": "Cluster recommendations",
+  "Cluster recommendation": "Cluster recommendation",
   "Cluster scope migrations": "Cluster scope migrations",
   "Cluster status": "Cluster status",
   "Cluster utilization": "Cluster utilization",
