+++ locales/es/plugin__kubevirt-plugin.json
@@ -17,11 +17,9 @@
   "{{availableText}} available": "{{availableText}} disponible",
   "{{badList}} is not a valid {{searchKey}} value.": "{{badList}} is not a valid {{searchKey}} value.",
   "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.": "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.",
+  "{{clusterCount}} clusters excluded": "{{clusterCount}} clusters excluded",
   "{{clustersCount}} clusters found": "{{clustersCount}} clústeres encontrados",
   "{{clustersCount}} clusters, {{projectsCount}} projects found": "{{clustersCount}} clústeres,{{projectsCount}} proyectos encontrados",
-  "{{count}} clusters excluded_one": "{{count}} clusters excluded",
-  "{{count}} clusters excluded_many": "{{count}} clusters excluded",
-  "{{count}} clusters excluded_other": "{{count}} clusters excluded",
   "{{count}} Expression_one": "{{count}} expresión_one",
   "{{count}} Expression_many": "{{count}} Expression",
   "{{count}} Expression_other": "{{count}} expresiones_other",
@@ -409,9 +407,11 @@
   "Clear search": "Borrar búsqueda",
   "Clear selection": "Clear selection",
   "Click \"Configure features\" to install this feature": "Haga clic en “Configurar funcionalidades” para instalar esta funcionalidad.",
+  "Click a cluster to select it as the migration target.": "Click a cluster to select it as the migration target.",
   "Click any example below to try it in the search bar": "Click any example below to try it in the search bar",
   "Click Connect to open serial console.": "Haga clic en Conectar para abrir la consola de serie.",
   "Click Connect to open the VNC console.": "Haga clic en Conectar para abrir la consola VNC.",
+  "Click the cluster to select it as the migration target.": "Click the cluster to select it as the migration target.",
   "Clone": "Clonar",
   "Clone {{kind}}": "Clonar {{kind}}",
   "Clone {{sourceKind}}": "Clonar {{sourceKind}}",
@@ -448,7 +448,7 @@
   "Cluster preferences": "Preferencias del clúster",
   "Cluster Preferences": "Preferencias del clúster",
   "Cluster preferences table": "Tabla de preferencias de clúster",
-  "Cluster recommendations": "Cluster recommendations",
+  "Cluster recommendation": "Cluster recommendation",
   "Cluster scope migrations": "Migraciones del alcance del clúster",
   "Cluster status": "Estado del clúster",
   "Cluster utilization": "Utilización del clúster",
