+++ locales/fr/plugin__kubevirt-plugin.json
@@ -17,11 +17,9 @@
   "{{availableText}} available": "{{availableText}} disponible(s)",
   "{{badList}} is not a valid {{searchKey}} value.": "{{badList}} is not a valid {{searchKey}} value.",
   "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.": "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.",
+  "{{clusterCount}} clusters excluded": "{{clusterCount}} clusters excluded",
   "{{clustersCount}} clusters found": "{{clustersCount}} groupes trouvés",
   "{{clustersCount}} clusters, {{projectsCount}} projects found": "{{clustersCount}} groupes,{{projectsCount}} projets trouvés",
-  "{{count}} clusters excluded_one": "{{count}} clusters excluded",
-  "{{count}} clusters excluded_many": "{{count}} clusters excluded",
-  "{{count}} clusters excluded_other": "{{count}} clusters excluded",
   "{{count}} Expression_one": "{{count}} Expression_une",
   "{{count}} Expression_many": "{{count}} Expression",
   "{{count}} Expression_other": "{{count}} Expression_autre",
@@ -409,9 +407,11 @@
   "Clear search": "Effacer la recherche",
   "Clear selection": "Clear selection",
   "Click \"Configure features\" to install this feature": "Cliquez sur «Configurer les fonctionnalités» pour installer cette fonctionnalité",
+  "Click a cluster to select it as the migration target.": "Click a cluster to select it as the migration target.",
   "Click any example below to try it in the search bar": "Click any example below to try it in the search bar",
   "Click Connect to open serial console.": "Cliquez sur Connecter pour ouvrir la console série.",
   "Click Connect to open the VNC console.": "Cliquez sur Connecter pour ouvrir la consoleVNC.",
+  "Click the cluster to select it as the migration target.": "Click the cluster to select it as the migration target.",
   "Clone": "Cloner",
   "Clone {{kind}}": "Cloner {{kind}}",
   "Clone {{sourceKind}}": "Cloner {{sourceKind}}",
@@ -448,7 +448,7 @@
   "Cluster preferences": "Préférences de cluster",
   "Cluster Preferences": "Préférences de cluster",
   "Cluster preferences table": "Tableau des préférences de cluster",
-  "Cluster recommendations": "Cluster recommendations",
+  "Cluster recommendation": "Cluster recommendation",
   "Cluster scope migrations": "Migrations de portée de cluster",
   "Cluster status": "Statut du cluster",
   "Cluster utilization": "Utilisation des clusters",
