+++ locales/ja/plugin__kubevirt-plugin.json
@@ -17,9 +17,9 @@
   "{{availableText}} available": "利用可能: {{availableText}}",
   "{{badList}} is not a valid {{searchKey}} value.": "{{badList}} is not a valid {{searchKey}} value.",
   "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.": "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.",
+  "{{clusterCount}} clusters excluded": "{{clusterCount}} clusters excluded",
   "{{clustersCount}} clusters found": "{{clustersCount}} 個のクラスターが見つかりました",
   "{{clustersCount}} clusters, {{projectsCount}} projects found": "{{clustersCount}} 個のクラスター、{{projectsCount}} 件のプロジェクトが見つかりました",
-  "{{count}} clusters excluded_other": "{{count}} clusters excluded",
   "{{count}} Expression_one": "{{count}} 式",
   "{{count}} Expression_other": "{{count}} 式",
   "{{count}} failed checks_one": "{{count}} 件のチェックに失敗しました",
@@ -396,9 +396,11 @@
   "Clear search": "検索をクリアする",
   "Clear selection": "Clear selection",
   "Click \"Configure features\" to install this feature": "この機能をインストールするには、「機能の設定」をクリックしてください",
+  "Click a cluster to select it as the migration target.": "Click a cluster to select it as the migration target.",
   "Click any example below to try it in the search bar": "Click any example below to try it in the search bar",
   "Click Connect to open serial console.": "「Connect」をクリックしてシリアルコンソールを開きます。",
   "Click Connect to open the VNC console.": "「Connect」をクリックして VNC コンソールを開きます。",
+  "Click the cluster to select it as the migration target.": "Click the cluster to select it as the migration target.",
   "Clone": "クローン",
   "Clone {{kind}}": "{{kind}} のクローン作成",
   "Clone {{sourceKind}}": "{{sourceKind}} のクローン作成",
@@ -435,7 +437,7 @@
   "Cluster preferences": "クラスター設定",
   "Cluster Preferences": "クラスター設定",
   "Cluster preferences table": "クラスター優先設定テーブル",
-  "Cluster recommendations": "Cluster recommendations",
+  "Cluster recommendation": "Cluster recommendation",
   "Cluster scope migrations": "クラスタースコープの移行",
   "Cluster status": "クラスターステータス",
   "Cluster utilization": "クラスターの使用率",
