+++ locales/ko/plugin__kubevirt-plugin.json
@@ -17,9 +17,9 @@
   "{{availableText}} available": "{{availableText}} 사용 가능",
   "{{badList}} is not a valid {{searchKey}} value.": "{{badList}} is not a valid {{searchKey}} value.",
   "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.": "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.",
+  "{{clusterCount}} clusters excluded": "{{clusterCount}} clusters excluded",
   "{{clustersCount}} clusters found": "{{clustersCount}} 클러스터가 발견됨",
   "{{clustersCount}} clusters, {{projectsCount}} projects found": "{{clustersCount}} 클러스터, {{projectsCount}} 프로젝트 발견",
-  "{{count}} clusters excluded_other": "{{count}} clusters excluded",
   "{{count}} Expression_one": "{{count}} 표현식",
   "{{count}} Expression_other": "{{count}} 표현식",
   "{{count}} failed checks_one": "{{count}} 검사 실패",
@@ -396,9 +396,11 @@
   "Clear search": "검색 지우기",
   "Clear selection": "Clear selection",
   "Click \"Configure features\" to install this feature": "이 기능을 설치하려면 \"Configure features\"를 클릭합니다.",
+  "Click a cluster to select it as the migration target.": "Click a cluster to select it as the migration target.",
   "Click any example below to try it in the search bar": "Click any example below to try it in the search bar",
   "Click Connect to open serial console.": "연결을 클릭하여 직렬 콘솔을 엽니다.",
   "Click Connect to open the VNC console.": "연결을 클릭하여 VNC 콘솔을 엽니다.",
+  "Click the cluster to select it as the migration target.": "Click the cluster to select it as the migration target.",
   "Clone": "복제",
   "Clone {{kind}}": "{{kind}} 복제",
   "Clone {{sourceKind}}": "{{sourceKind}} 복제",
@@ -435,7 +437,7 @@
   "Cluster preferences": "클러스터 기본 설정",
   "Cluster Preferences": "클러스터 기본 설정",
   "Cluster preferences table": "클러스터 기본 설정 테이블",
-  "Cluster recommendations": "Cluster recommendations",
+  "Cluster recommendation": "Cluster recommendation",
   "Cluster scope migrations": "클러스터 범위 마이그레이션",
   "Cluster status": "클러스터 상태",
   "Cluster utilization": "클러스터 활용",
