+++ locales/zh/plugin__kubevirt-plugin.json
@@ -17,9 +17,9 @@
   "{{availableText}} available": "{{availableText}} 可用",
   "{{badList}} is not a valid {{searchKey}} value.": "{{badList}} is not a valid {{searchKey}} value.",
   "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.": "{{badList}} is not a valid {{searchKey}} value. Your search uses {{validList}}.",
+  "{{clusterCount}} clusters excluded": "{{clusterCount}} clusters excluded",
   "{{clustersCount}} clusters found": "找到 {{clustersCount}} 个集群",
   "{{clustersCount}} clusters, {{projectsCount}} projects found": "找到 {{clustersCount}} 个集群，{{projectsCount}} 个项目",
-  "{{count}} clusters excluded_other": "{{count}} clusters excluded",
   "{{count}} Expression_one": "{{count}} 表达式",
   "{{count}} Expression_other": "{{count}} 表达式",
   "{{count}} failed checks_one": "{{count}} 个失败的检查",
@@ -396,9 +396,11 @@
   "Clear search": "清除搜索",
   "Clear selection": "Clear selection",
   "Click \"Configure features\" to install this feature": "点 \"Configure features\" 以安装此功能",
+  "Click a cluster to select it as the migration target.": "Click a cluster to select it as the migration target.",
   "Click any example below to try it in the search bar": "Click any example below to try it in the search bar",
   "Click Connect to open serial console.": "点 Connect 打开串行控制台。",
   "Click Connect to open the VNC console.": "点 Connect 打开 VNC 控制台。",
+  "Click the cluster to select it as the migration target.": "Click the cluster to select it as the migration target.",
   "Clone": "克隆",
   "Clone {{kind}}": "克隆 {{kind}}",
   "Clone {{sourceKind}}": "克隆 {{sourceKind}}",
@@ -435,7 +437,7 @@
   "Cluster preferences": "集群首选",
   "Cluster Preferences": "集群首选",
   "Cluster preferences table": "集群首选项表",
-  "Cluster recommendations": "Cluster recommendations",
+  "Cluster recommendation": "Cluster recommendation",
   "Cluster scope migrations": "集群范围迁移",
   "Cluster status": "集群状态",
   "Cluster utilization": "集群利用率",
