--- src/multicluster/components/CrossClusterMigration/components/ClusterRecommendationPanel.scss
@@ -1,42 +1,0 @@
-.cluster-recommendation-panel {
-  margin-top: var(--pf-t--global--spacer--md);
-  display: flex;
-  flex-direction: column;
-  gap: var(--pf-t--global--spacer--sm);
-
-  &--loading {
-    display: flex;
-    align-items: center;
-    gap: var(--pf-t--global--spacer--sm);
-    padding: var(--pf-t--global--spacer--md) 0;
-  }
-
-  &__list {
-    display: flex;
-    flex-direction: column;
-    gap: var(--pf-t--global--spacer--xs);
-  }
-
-  &__candidate {
-    border: var(--pf-t--global--border--width--regular) solid
-      var(--pf-t--global--border--color--default);
-    border-radius: var(--pf-t--global--border--radius--medium);
-    padding: var(--pf-t--global--spacer--sm) var(--pf-t--global--spacer--md);
-    cursor: pointer;
-    transition: background-color 0.15s ease;
-
-    &:hover {
-      background-color: var(--pf-t--global--background--color--primary--hover);
-    }
-  }
-
-  &__scores {
-    display: flex;
-    gap: var(--pf-t--global--spacer--md);
-    flex-wrap: wrap;
-  }
-
-  &__top-label {
-    margin-left: var(--pf-t--global--spacer--sm);
-  }
-}
